requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.7.1
dj-database-url==2.1.0
dj-email-url==1.0.6
Django==5.0.1
django-allauth==0.60.0
django-cache-url==3.4.5
django-cors-headers==4.3.1
environs==10.0.0
idna==3.6
marshmallow==3.20.1
oauthlib==3.2.2
packaging==23.2
psycopg==3.1.16
psycopg-binary==3.1.16
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
whitenoise==6.6.0
pillow==10.2.0
celery==5.3.6
redis==5.0.1
